refactoring in proces

+++ app/src/main/java/ru/butterbean/easyrent/screens/room/item_models/ArchiveReserveModel.kt
@@ -8,7 +8,7 @@ import ru.butterbean.easyrent.utils.APP_ACTIVITY
 
 class ArchiveReserveModel(
     val roomId: Long,
-    private val archivedCount: Int
+    private val archivedCount: Long
 ) : CommonReserveModel {
     override fun getItemViewType(): Int = CommonReserveModel.ARCHIVE
